requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pillow==10.0.1
lxml==4.9.3
webdriver-manager==4.0.1
